requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
Flask-Mail==0.9.1
twilio==7.16.0